09_vm_stockpd_linebot/requirements.txt
google-cloud-storage
google-cloud-firestore
flask
gunicorn
line-bot-sdk
google-cloud-logging
Pandas
yfinance
mpl_finance
matplotlib